Fully merge study items
