+++ src/assets/style/navigation-drawers/study-item.css.less
@@ -13,6 +13,7 @@
   white-space: nowrap;
   overflow: hidden;
   text-overflow: ellipsis;
+  margin-left: 0;
 }
 
 .error-item {
